storage management: move atomWithAsyncStorage to storage.ts and update docs
