+++ atoms/theme-atoms.ts
@@ -1,34 +1,12 @@
-import AsyncStorage from '@react-native-async-storage/async-storage';
 import { atom } from 'jotai';
 import { appThemeDark, appThemeLight } from '../lib/theme';
-import { STORAGE_KEYS } from './storage';
+import { atomWithAsyncStorage, STORAGE_KEYS } from './storage';
 
 export type ThemeMode = "auto" | "light" | "dark";
 
-const atomWithAsyncStorage = (key: string, initialValue: any) => {
-  const baseAtom = atom(initialValue)
-  baseAtom.onMount = (setValue) => {
-    ;(async () => {
-      const item = await AsyncStorage.getItem(key)
-      if (item !== null) {
-        setValue(JSON.parse(item))
-      }
-    })()
-  }
-  const derivedAtom = atom(
-    (get) => get(baseAtom),
-    (get, set, update) => {
-      const nextValue =
-        typeof update === 'function' ? update(get(baseAtom)) : update
-      set(baseAtom, nextValue)
-      AsyncStorage.setItem(key, JSON.stringify(nextValue))
-    },
-  )
-  return derivedAtom
-}
 
 // Base atom for theme mode - kept up to date in the ThemeProvider
-export const themeModeAtom = atomWithAsyncStorage(
+export const themeModeAtom = atomWithAsyncStorage<ThemeMode>(
   STORAGE_KEYS.THEME_MODE, 
   "auto"
 );
